Tiktokenizer Test Results / Explanation
First off let's talk about Tokenizers. Tokenizers are essential to a LLM, as it basically is what translate raw text input into a format that the LLM can understand. Converting this raw text into numbers is a process that has many layers. Tokenizers allow you to train and input larger and larger amounts in of data into LLM's with maximum efficiency and compression.

When training an LLM with these massive amounts of data, it's important to stay efficient, and have the LLM allocate it's resources to where they need to be used; This is where Mixture of Experts come in. Mixture of Experts decides what resources in the LLM are best suited to handle the task given to them.

So last but not least, Coding Assistance. LLM's are probably the absolute best tool regarding assistance with coding. You can do a multitude of different things with LLM's when coding; For example, you could feed it a chunk of complex code and ask for it to explain it to you in simple terms, you could use it to debug your code, you can ask it to generate code based on your input, and much more. LLM's are far from perfect, however they are pretty dang powerful, and it's not hard to imagine the future possibility.